POC: Forked eslint-scope as dependency
