+++ lib/utils.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var t = require("babel-types");
+var eslintScope = require("eslint-scope");
 
 /**
  * Exported utilities
@@ -9,6 +10,8 @@ module.exports = {
   visitDecorators,
   visitTypeAnnotation,
   checkIdentifierOrVisit,
+  createScopeVariable,
+  nestTypeParamScope,
 };
 
 var propertyTypes = {
@@ -117,4 +120,37 @@ function checkIdentifierOrVisit(node) {
   } else {
     visitTypeAnnotation.call(this, node);
   }
+}
+
+function createScopeVariable(node, name) {
+  this.currentScope().variableScope.__define(
+    name,
+    new eslintScope.Definition("Variable", name, node, null, null, null)
+  );
+}
+
+function nestTypeParamScope(manager, node) {
+  var parentScope = manager.__currentScope;
+  var scope = new eslintScope.Scope(
+    manager,
+    "type-parameters",
+    parentScope,
+    node,
+    false
+  );
+  manager.__nestScope(scope);
+  for (var j = 0; j < node.typeParameters.params.length; j++) {
+    var name = node.typeParameters.params[j];
+    scope.__define(
+      name,
+      new eslintScope.Definition("TypeParameter", name, name)
+    );
+    if (name.typeAnnotation) {
+      checkIdentifierOrVisit.call(this, name);
+    }
+  }
+  scope.__define = function() {
+    return parentScope.__define.apply(parentScope, arguments);
+  };
+  return scope;
 }
