+++ package.json
@@ -12,10 +12,11 @@
     "url": "https://github.com/babel/babel-eslint.git"
   },
   "dependencies": {
-    "babel-code-frame": "7.0.0-alpha.17",
-    "babel-traverse": "7.0.0-alpha.17",
-    "babel-types": "7.0.0-alpha.17",
-    "babylon": "7.0.0-beta.18"
+    "babel-code-frame": "7.0.0-alpha.19",
+    "babel-traverse": "7.0.0-alpha.19",
+    "babel-types": "7.0.0-alpha.19",
+    "babylon": "7.0.0-beta.22",
+    "eslint-scope": "git://github.com/eslint/eslint-scope.git#4f0f8c878fd9a2440e1c6c3e05b6981a9494a156"
   },
   "scripts": {
     "test": "npm run lint && npm run test-only",
